feat: refactored the code for more standard structure.
Add: Redux for state management
