+++ client/src/hooks/useIntersectionObserver.js
@@ -1,0 +1,23 @@
+import { useState, useEffect } from "react";
+
+export const useIntersectionObserver = (ref, options) => {
+  const [isVisible, setIsVisible] = useState(false);
+
+  useEffect(() => {
+    const observer = new IntersectionObserver(([entry]) => {
+      setIsVisible(entry.isIntersecting);
+    }, options);
+
+    if (ref.current) {
+      observer.observe(ref.current);
+    }
+
+    return () => {
+      if (ref.current) {
+        observer.unobserve(ref.current);
+      }
+    };
+  }, [ref, options]);
+
+  return isVisible;
+};
