--- client/src/pages/AboutPage.jsx
@@ -1,5 +1,0 @@
-const AboutPage = () => {
-  return <div>AboutPage</div>;
-};
-
-export default AboutPage;
