--- client/src/pages/AuthPage.jsx
@@ -1,5 +1,0 @@
-const AuthPage = () => {
-  return <div>AuthPage</div>;
-};
-
-export default AuthPage;
