--- client/src/pages/ContactPage.jsx
@@ -1,5 +1,0 @@
-const ContactPage = () => {
-  return <div>ContactPage</div>;
-};
-
-export default ContactPage;
