--- client/src/pages/HomePage.jsx
@@ -1,15 +1,0 @@
-import HeroSection from "../components/home/HeroSection";
-import ServiceCategories from "../components/home/ServiceCategories";
-import FeaturedVendors from "../components/home/FeaturedVendors";
-
-const HomePage = () => {
-  return (
-    <>
-      <HeroSection />
-      <ServiceCategories />
-      <FeaturedVendors />
-    </>
-  );
-};
-
-export default HomePage;
