--- client/src/pages/ServicesPage.jsx
@@ -1,5 +1,0 @@
-const ServicesPage = () => {
-  return <div>ServicesPage</div>;
-};
-
-export default ServicesPage;
