+++ client/src/pages/homepage/HomePage.jsx
@@ -1,0 +1,15 @@
+import HeroSection from '../../components/sections/home/HeroSection';
+import ServiceCategories from '../../components/sections/home/ServiceCategories';
+import VendorsSection from '../../components/sections/home/VendorsSection';
+
+const HomePage = () => {
+    return (
+        <main>
+            <HeroSection />
+            <ServiceCategories />
+            <VendorsSection />
+        </main>
+    );
+};
+
+export default HomePage;
