+++ client/src/store/store.js
@@ -1,0 +1,8 @@
+import { configureStore } from "@reduxjs/toolkit";
+import uiReducer from "./ui/uiSlice";
+
+export const store = configureStore({
+  reducer: {
+    ui: uiReducer,
+  },
+});
