+++ client/src/utils/constants/contactInfo.js
@@ -1,0 +1,7 @@
+export const CONTACT_INFO = [
+  { icon: "🏢", text: "AnyJob Inc." },
+  { icon: "📍", text: "123 Service Street" },
+  { icon: "🌆", text: "Marketplace City, MP 12345" },
+  { icon: "✉️", text: "info@anyjob.com" },
+  { icon: "📱", text: "(123) 456-7890" },
+];
