+++ client/src/utils/constants/featureBadges.js
@@ -1,0 +1,5 @@
+export const FEATURE_BADGES = [
+  { icon: "⚡", text: "Fast & Reliable" },
+  { icon: "🛡️", text: "Secure Payments" },
+  { icon: "🔍", text: "Verified Providers" },
+];
