+++ client/src/utils/constants/languages.js
@@ -1,0 +1,5 @@
+export const LANGUAGES = [
+  { code: "en", name: "English" },
+  { code: "es", name: "Español" },
+  { code: "fr", name: "Français" },
+];
