+++ client/src/utils/constants/legalLinks.js
@@ -1,0 +1,5 @@
+export const LEGAL_LINKS = [
+  { name: "Privacy Policy", path: "/privacy" },
+  { name: "Terms of Service", path: "/terms" },
+  { name: "Cookie Policy", path: "/cookies" },
+];
