+++ client/src/utils/constants/navItems.js
@@ -1,0 +1,6 @@
+export const NAV_ITEMS = [
+  { name: "Home", path: "/" },
+  { name: "Services", path: "/services" },
+  { name: "About", path: "/about" },
+  { name: "Contact", path: "/contact" },
+];
