+++ client/src/utils/constants/socialLinks.js
@@ -1,24 +1,4 @@
-export const POPULAR_CATEGORIES = [
-  { name: "Plumbing", icon: "🔧" },
-  { name: "Cleaning", icon: "🧹" },
-  { name: "Electrical", icon: "💡" },
-  { name: "Moving", icon: "📦" },
-];
-
-export const FEATURE_BADGES = [
-  { icon: "⚡", text: "Fast & Reliable" },
-  { icon: "🛡️", text: "Secure Payments" },
-  { icon: "🔍", text: "Verified Providers" },
-];
-
-export const NAV_ITEMS = [
-  { name: "Home", path: "/" },
-  { name: "Services", path: "/services" },
-  { name: "About", path: "/about" },
-  { name: "Contact", path: "/contact" },
-];
-
-export const SOCIAL_PLATFORMS = [
+export const SOCIAL_LINKS = [
   {
     name: "Twitter",
     url: "https://twitter.com",
@@ -39,24 +19,4 @@ export const SOCIAL_PLATFORMS = [
     url: "https://linkedin.com",
     icon: "M19 0h-14c-2.761 0-5 2.239-5 5v14c0 2.761 2.239 5 5 5h14c2.762 0 5-2.239 5-5v-14c0-2.761-2.238-5-5-5zm-11 19h-3v-11h3v11zm-1.5-12.268c-.966 0-1.75-.79-1.75-1.764s.784-1.764 1.75-1.764 1.75.79 1.75 1.764-.783 1.764-1.75 1.764zm13.5 12.268h-3v-5.604c0-3.368-4-3.113-4 0v5.604h-3v-11h3v1.765c1.396-2.586 7-2.777 7 2.476v6.759z",
   },
-];
-
-export const LEGAL_LINKS = [
-  { name: "Privacy Policy", path: "/privacypolicy" },
-  { name: "Terms of Service", path: "/termsofservice" },
-  { name: "Cookie Policy", path: "/cookiepolicy" },
-];
-
-export const CONTACT_INFO = [
-  { icon: "🏢", text: "AnyJob Inc." },
-  { icon: "📍", text: "123 Service Street" },
-  { icon: "🌆", text: "Marketplace City, MP 12345" },
-  { icon: "✉️", text: "info@anyjob.com" },
-  { icon: "📱", text: "(123) 456-7890" },
-];
-
-export const LANGUAGES = [
-  { code: "en", name: "English" },
-  { code: "es", name: "Español" },
-  { code: "fr", name: "Français" },
 ];
