+++ client/src/utils/data/vendors.js
@@ -1,0 +1,34 @@
+export const vendors = [
+  {
+    id: 1,
+    name: "John Doe",
+    image: "https://images.unsplash.com/photo-1560250097-0b93528c311a",
+    rating: 4.8,
+    service: "Professional Plumber",
+    price: "From $50/hr",
+  },
+  {
+    id: 2,
+    name: "Jane Smith",
+    image: "https://images.unsplash.com/photo-1573496359142-b8d87734a5a2",
+    rating: 4.9,
+    service: "House Cleaning",
+    price: "From $30/hr",
+  },
+  {
+    id: 3,
+    name: "Mike Johnson",
+    image: "https://images.unsplash.com/photo-1566753323558-f4e0952af115",
+    rating: 4.7,
+    service: "Electrician",
+    price: "From $55/hr",
+  },
+  {
+    id: 4,
+    name: "Sarah Williams",
+    image: "https://images.unsplash.com/photo-1580489944761-15a19d654956",
+    rating: 5.0,
+    service: "Math Tutor",
+    price: "From $25/hr",
+  },
+];
