+++ package.json
@@ -23,6 +23,7 @@
   "author": "Avishka Gihan",
   "license": "MIT",
   "dependencies": {
+    "anyjob-platform": "file:",
     "bcryptjs": "^3.0.2",
     "cors": "^2.8.5",
     "dotenv": "^16.4.7",
